Pre: Move PushComputed up module hierarchy; make it public.
